Improve organization of tup-hooks by grouping hooks and types according to business domain.
