+++ libs/tup-hooks/README.md
@@ -4,4 +4,4 @@ This library was generated with [Nx](https://nx.dev).
 
 ## Running unit tests
 
-Run `nx test tup-hooks` to execute the unit tests via [Jest](https://jestjs.io).
+Run `nx test tup-hooks` to execute the unit tests via [Vitest](https://vitest.dev/).
