+++ libs/tup-hooks/src/auth/useProfile.ts
@@ -1,6 +1,6 @@
 import { UseQueryResult } from '@tanstack/react-query';
 import { UserProfile } from '.';
-import { useGet } from './requests';
+import { useGet } from '../requests';
 
 // Query to retrieve the user's profile object.
 const useProfile = (): UseQueryResult<UserProfile> => {
