+++ libs/tup-hooks/src/mfa/index.ts
@@ -1,0 +1,46 @@
+export type MfaTokenResponse = {
+  token?: {
+    active: boolean;
+    description: string;
+    id: number;
+    locked: boolean;
+    revoked: boolean;
+    serial: string;
+    tokentype: 'sms' | 'totp';
+    user_id: string;
+    user_realm: string;
+    username: string;
+    rollout_state: 'verify' | 'enrolled';
+  };
+};
+
+export type MfaPairingResponse = {
+  googleurl: {
+    description: string;
+    img: string;
+  };
+  rollout_state: string;
+  serial: string;
+};
+
+export type MfaValidationResponse = {
+  detail: {
+    otplen: number;
+    serial: string;
+    type: string;
+    message: string;
+  };
+  result: {
+    authentication: 'ACCEPT' | 'REJECT';
+    status: boolean;
+    value: boolean;
+  };
+};
+
+export {
+  useMfa,
+  useMfaPairTotp,
+  useMfaPairSms,
+  useMfaVerify,
+  useMfaDelete,
+} from './useMfa';
