+++ libs/tup-hooks/src/news/index.ts
@@ -1,0 +1,16 @@
+export type UserNewsResponse = {
+  ID: string;
+  Updates: {
+    AnnouncementUpdate: { ID: string; PostedDate: string; Content: string }[];
+  };
+  Author: string;
+  PostedDate: string;
+  AnnouncementDate: string;
+  ArchiveDate: string;
+  Title: string;
+  Subtitle: string;
+  WebTitle: string;
+  Content: string;
+};
+
+export { default as useUserNews } from './useUserNews';
