+++ libs/tup-hooks/src/news/useUserNews.tsx
@@ -1,6 +1,8 @@
-import { useGet } from './requests';
+import { useGet } from '../requests';
 import { UserNewsResponse } from '.';
 
-export const useUserNews = () => {
+const useUserNews = () => {
   return useGet<UserNewsResponse[]>({ endpoint: '/news', key: ['news'] });
 };
+
+export default useUserNews;
