+++ libs/tup-hooks/src/projects/useProjectScienceField.ts
@@ -1,6 +1,6 @@
 import { UseQueryResult } from '@tanstack/react-query';
 import { ProjectFieldOfScience } from '.';
-import { useGet } from './requests';
+import { useGet } from '../requests';
 
 // Query to retrieve the user's active projects.
 const useProjectScienceField = (): UseQueryResult<ProjectFieldOfScience[]> => {
