+++ tsconfig.base.json
@@ -20,7 +20,6 @@
       "@tacc/core-wrappers": ["libs/core-wrappers/src/index.ts"],
       "@tacc/tup-components": ["libs/tup-components/src/index.ts"],
       "@tacc/tup-hooks": ["libs/tup-hooks/src/index.ts"],
-      "@tacc/tup-hooks/*": ["libs/tup-hooks/src/*"],
       "@tacc/tup-testing": ["libs/tup-testing/src/index.ts"],
       "@tacc/tup-ui/*": ["apps/tup-ui/src/*"]
     }
